Refactor webhooks and add queue event webhook
