+++ src/pages/account/queue-event-webhook/delete.js
@@ -1,0 +1,10 @@
+import DeleteWebhook from '../../../components/webhooks/DeleteWebhook';
+
+const QueueEventWebhookDelete = () => {
+  return <DeleteWebhook
+    hookName="Queue Event"
+    hookProp="queue_event_hook"
+  />;
+};
+
+export default QueueEventWebhookDelete;
